+++ NMJToolBox/KevinCastejon/EditorToolbox/Modules/Editor/Attributes/HideOnPrefabDrawer.cs
@@ -1,5 +1,5 @@
 using UnityEditor;
-using UnityEditor.Experimental.SceneManagement;
+
 using UnityEngine;
 
 namespace NMJToolBox
@@ -10,14 +10,14 @@ namespace NMJToolBox
         public override float GetPropertyHeight(SerializedProperty property, GUIContent label)
         {
             HideOnPrefabAttribute att = (HideOnPrefabAttribute)attribute;
-            bool hidden = att.invert ? PrefabStageUtility.GetCurrentPrefabStage() == null : PrefabStageUtility.GetCurrentPrefabStage() != null;
+            bool hidden = att.invert ? UnityEditor.SceneManagement.PrefabStageUtility.GetCurrentPrefabStage() == null : UnityEditor.SceneManagement.PrefabStageUtility.GetCurrentPrefabStage() != null;
             return hidden ? 0 : EditorGUI.GetPropertyHeight(property, label);
         }
 
         public override void OnGUI(Rect position, SerializedProperty property, GUIContent label)
         {
             HideOnPrefabAttribute att = (HideOnPrefabAttribute)attribute;
-            bool hidden = att.invert ? PrefabStageUtility.GetCurrentPrefabStage() == null : PrefabStageUtility.GetCurrentPrefabStage() != null;
+            bool hidden = att.invert ? UnityEditor.SceneManagement.PrefabStageUtility.GetCurrentPrefabStage() == null : UnityEditor.SceneManagement.PrefabStageUtility.GetCurrentPrefabStage() != null;
             if (hidden) return;
             EditorGUI.PropertyField(position, property, label);
         }