+++ CHANGELOG.md
@@ -1,15 +1,28 @@
 # NMJToolKit version history
+---
+## v1.0.3
+  Changes:
+    -Added JsonParser from string to generic [Serializable] object
+    -Added String Trim Extension for start and end
+    -Added AssetDatabaseExtensions ande AssetDatabaseUtilities
+    -Added CustomListView, Editor only use
+
+---
+## v1.0.2
+  Changes:
+    -Deleted Hotkey Scenes Switcher from https://github.com/vimpsoft/ScenesSwitcher by Andrey Tsarkov
+    -Deleted Hierarchy 2 from https://github.com/truongnguyentungduy/hierarchy-2 by truongnguyentungduy
 
 ---
 
 ## v1.0.1
- Fix:
+ Changes:
   - Added Hotkey Scenes Switcher from https://github.com/vimpsoft/ScenesSwitcher by Andrey Tsarkov
 
 ---
 
 ## v1.0.1
- Fix:
+ Changes:
   - Added Hierarchy 2 from https://github.com/truongnguyentungduy/hierarchy-2 by truongnguyentungduy
 
 ---